+++ src/main/java/com/gmail/goosius/siegewar/utils/SiegeWarDistanceUtil.java
@@ -224,4 +224,26 @@ public class SiegeWarDistanceUtil {
 		Location location2 = siege.getFlagLocation();
 		return (int)MathUtil.distance(location1.getX(), location2.getX(), location1.getZ(), location2.getZ());
 	}
+
+	/**
+	 * Determine if the target location is protected by trap warfare mitigation
+	 *
+	 * @param targetLocation target location
+	 * @param siegeBannerLocation location of nearby siege banner
+	 * @param protectionRadiusBlocks protection radius in blocks
+	 * @param isProtectionBelowBannerOnly true if we should protect only below banner
+	 *
+	 * @return true if the location is protected
+	 */
+	public static boolean isTargetLocationProtectedByTrapWarfareMitigation(Location targetLocation, Location siegeBannerLocation, int protectionRadiusBlocks, boolean isProtectionBelowBannerOnly) {
+		if(!TownyAPI.getInstance().isWilderness(targetLocation)) {
+			return false;  //In town. Protection does not apply.
+		} else if(isProtectionBelowBannerOnly && targetLocation.getY() >= siegeBannerLocation.getY()) {
+			return false;  //Above banner. Protection does not apply.
+		} else if(areLocationsCloseHorizontally(targetLocation, siegeBannerLocation, protectionRadiusBlocks)) {
+			return true;   //Target location is protected
+		} else {
+			return false;  //Target location is not protected
+		}
+	}
 }
